changed structure of code and add logic of correct saving products in localStorage
