+++ src/redux/slices/filter/selectors.ts
@@ -1,0 +1,3 @@
+import { RootState } from "../../store";
+
+export const selectFilter = (state: RootState) => state.filter;
