+++ src/redux/slices/products/selectors.ts
@@ -1,0 +1,3 @@
+import { RootState } from "../../store";
+
+export const selectProducts = (state: RootState) => state.product;
