+++ src/redux/store.ts
@@ -1,8 +1,8 @@
 import { useDispatch } from "react-redux";
 import { configureStore } from "@reduxjs/toolkit";
-import filterReducer from "./slices/filterSlice";
-import cartReducer from "./slices/cartSlice";
-import productsReducer from "./slices/productsSlice";
+import filterReducer from "./slices/filter/slice";
+import cartReducer from "./slices/cart/slice";
+import productsReducer from "./slices/products/slice";
 
 export const store = configureStore({
   reducer: {
